# docs/contract/openapi.json
{
  "components": {
    "schemas": {
      "AbstainView": {
        "properties": {
          "action": {
            "type": "string"
          },
          "p_fail": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "action",
          "p_fail"
        ],
        "type": "object"
      },
      "AnalysisView": {
        "properties": {
          "intent": {
            "type": "string"
          },
          "out_of_corpus_terms": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "primary": {
            "type": "object"
          },
          "score": {
            "type": "number"
          }
        },
        "required": [
          "intent",
          "primary",
          "score"
        ],
        "type": "object"
      },
      "Calibration": {
        "properties": {
          "program": {
            "type": "string"
          },
          "threshold": {
            "type": "number"
          },
          "threshold_version": {
            "type": "string"
          },
          "tier": {
            "type": "string"
          }
        },
        "required": [
          "program",
          "threshold",
          "threshold_version",
          "tier"
        ],
        "type": "object"
      },
      "CommittedView": {
        "properties": {
          "at": {
            "type": "object"
          },
          "belief_version": {
            "type": "string"
          },
          "calibration": {
            "$ref": "#/components/schemas/Calibration"
          },
          "config_version": {
            "type": "string"
          },
          "corpus_version": {
            "type": "string"
          },
          "flow": {
            "type": "string"
          },
          "realm": {
            "type": "string"
          },
          "strategy_version": {
            "type": "string"
          }
        },
        "required": [
          "at",
          "belief_version",
          "calibration",
          "config_version",
          "corpus_version",
          "flow",
          "realm",
          "strategy_version"
        ],
        "type": "object"
      },
      "ConflictView": {
        "properties": {
          "fact_id": {
            "type": "string"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "values": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "fact_id",
          "source_ids",
          "values"
        ],
        "type": "object"
      },
      "CoverageView": {
        "properties": {
          "out_of_corpus_terms": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "value": {
            "type": "number"
          }
        },
        "required": [
          "value"
        ],
        "type": "object"
      },
      "DocResponse": {
        "properties": {
          "body": {
            "type": "string"
          },
          "encoding": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "span_end": {
            "type": "integer"
          },
          "span_start": {
            "type": "integer"
          },
          "src_bytes": {
            "type": "integer"
          },
          "src_digest": {
            "type": "string"
          }
        },
        "required": [
          "body",
          "id",
          "span_end",
          "span_start"
        ],
        "type": "object"
      },
      "EscalationRecord": {
        "properties": {
          "after": {
            "type": "string"
          },
          "before": {
            "type": "string"
          },
          "executed": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          },
          "triggered": {
            "type": "boolean"
          },
          "windows": {
            "type": "integer"
          }
        },
        "required": [
          "after",
          "before",
          "executed",
          "triggered",
          "windows"
        ],
        "type": "object"
      },
      "EvictionView": {
        "properties": {
          "dropped": {
            "type": "integer"
          },
          "merged": {
            "type": "integer"
          }
        },
        "required": [
          "dropped",
          "merged"
        ],
        "type": "object"
      },
      "FactView": {
        "properties": {
          "covered": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "judge": {
            "type": "string"
          },
          "near_miss": {
            "type": "number"
          },
          "query": {
            "type": "string"
          },
          "supports": {
            "items": {
              "$ref": "#/components/schemas/SupportView"
            },
            "type": "array"
          }
        },
        "required": [
          "covered",
          "id",
          "query"
        ],
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "corpus_docs": {
            "type": "integer"
          },
          "corpus_version": {
            "type": "string"
          },
          "realm": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "corpus_docs",
          "corpus_version",
          "realm",
          "status"
        ],
        "type": "object"
      },
      "IngestRequest": {
        "properties": {
          "body": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "body",
          "url"
        ],
        "type": "object"
      },
      "IngestResponse": {
        "properties": {
          "bytes": {
            "type": "integer"
          },
          "corpus_docs": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "bytes",
          "corpus_docs",
          "id"
        ],
        "type": "object"
      },
      "PriorView": {
        "properties": {
          "doc_id": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "signals": {
            "type": "object"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "doc_id",
          "score",
          "signals"
        ],
        "type": "object"
      },
      "QARequest": {
        "properties": {
          "question": {
            "type": "string"
          },
          "session": {
            "type": "string"
          }
        },
        "required": [
          "question",
          "session"
        ],
        "type": "object"
      },
      "QAResponse": {
        "properties": {
          "abstain": {
            "$ref": "#/components/schemas/AbstainView"
          },
          "analysis": {
            "$ref": "#/components/schemas/AnalysisView"
          },
          "answer": {
            "type": "string"
          },
          "citations": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "committed": {
            "$ref": "#/components/schemas/CommittedView"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/ConflictView"
            },
            "type": "array"
          },
          "coverage": {
            "$ref": "#/components/schemas/CoverageView"
          },
          "escalation": {
            "$ref": "#/components/schemas/EscalationRecord"
          },
          "eviction": {
            "$ref": "#/components/schemas/EvictionView"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/FactView"
            },
            "type": "array"
          },
          "prior": {
            "items": {
              "$ref": "#/components/schemas/PriorView"
            },
            "type": "array"
          },
          "question": {
            "type": "string"
          },
          "refusal_reason": {
            "type": "string"
          },
          "refused": {
            "type": "boolean"
          },
          "rerank": {
            "$ref": "#/components/schemas/RerankState"
          },
          "reuse": {
            "$ref": "#/components/schemas/ReuseState"
          },
          "route": {
            "$ref": "#/components/schemas/RouteView"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageView"
          },
          "windows": {
            "items": {
              "$ref": "#/components/schemas/WindowView"
            },
            "type": "array"
          }
        },
        "required": [
          "analysis",
          "answer",
          "citations",
          "committed",
          "coverage",
          "escalation",
          "eviction",
          "question",
          "refused",
          "rerank",
          "reuse",
          "route",
          "usage",
          "windows"
        ],
        "type": "object"
      },
      "RerankState": {
        "properties": {
          "applied": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "applied"
        ],
        "type": "object"
      },
      "ReuseState": {
        "properties": {
          "hit": {
            "type": "boolean"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "hit"
        ],
        "type": "object"
      },
      "RouteSignals": {
        "properties": {
          "calibration_program": {
            "type": "string"
          },
          "confidence": {
            "type": "number"
          },
          "confidence_known": {
            "type": "boolean"
          },
          "conflicts": {
            "type": "integer"
          },
          "coverage": {
            "type": "number"
          },
          "dead_rate": {
            "type": "number"
          },
          "facts_covered": {
            "type": "integer"
          },
          "facts_k": {
            "type": "integer"
          },
          "facts_missing": {
            "type": "integer"
          },
          "gap_thin": {
            "type": "boolean"
          },
          "margin": {
            "type": "number"
          },
          "threshold": {
            "type": "number"
          },
          "threshold_version": {
            "type": "string"
          },
          "tier": {
            "type": "string"
          },
          "windows": {
            "type": "integer"
          }
        },
        "required": [
          "calibration_program",
          "confidence",
          "confidence_known",
          "conflicts",
          "coverage",
          "dead_rate",
          "facts_covered",
          "facts_k",
          "facts_missing",
          "gap_thin",
          "margin",
          "threshold",
          "tier",
          "windows"
        ],
        "type": "object"
      },
      "RouteView": {
        "properties": {
          "action": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "signals": {
            "$ref": "#/components/schemas/RouteSignals"
          }
        },
        "required": [
          "action",
          "reason",
          "signals"
        ],
        "type": "object"
      },
      "StatusResponse": {
        "properties": {
          "embedder": {
            "type": "boolean"
          },
          "escalate": {
            "type": "boolean"
          },
          "reuse": {
            "type": "boolean"
          },
          "synthesis": {
            "type": "boolean"
          }
        },
        "required": [
          "embedder",
          "escalate",
          "reuse",
          "synthesis"
        ],
        "type": "object"
      },
      "SupportView": {
        "properties": {
          "score": {
            "type": "number"
          },
          "source_id": {
            "type": "string"
          },
          "span": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "score",
          "source_id",
          "span"
        ],
        "type": "object"
      },
      "UsageView": {
        "properties": {
          "completion_tokens": {
            "type": "integer"
          },
          "cost_known": {
            "type": "boolean"
          },
          "prompt_tokens": {
            "type": "integer"
          }
        },
        "required": [
          "completion_tokens",
          "cost_known",
          "prompt_tokens"
        ],
        "type": "object"
      },
      "WindowView": {
        "properties": {
          "score": {
            "type": "number"
          },
          "source_id": {
            "type": "string"
          },
          "span": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "score",
          "source_id",
          "span",
          "text",
          "title"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "从 Go 结构体反射生成；契约的裁判是代码，不是本文档。改形状先改代码。",
    "title": "cumulus qa api",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/v1/doc/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "span",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocResponse"
                }
              }
            },
            "description": "ok"
          },
          "404": {
            "description": "文档不存在"
          },
          "default": {
            "description": "error"
          }
        },
        "summary": "文档全文（点击引用看原文；span 参数带出高亮区间）"
      }
    },
    "/v1/health": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "description": "error"
          }
        },
        "summary": "探活：语料量与 realm"
      }
    },
    "/v1/ingest": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "ok"
          },
          "400": {
            "description": "空内容或坏 JSON"
          },
          "502": {
            "description": "链接取回失败"
          },
          "default": {
            "description": "error"
          }
        },
        "summary": "摄入：粘贴文本或给链接，入库后索引热重建"
      }
    },
    "/v1/qa": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QARequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAResponse"
                }
              }
            },
            "description": "ok"
          },
          "400": {
            "description": "请求不是合法 JSON 或 question 为空"
          },
          "500": {
            "description": "流程失败"
          },
          "default": {
            "description": "error"
          }
        },
        "summary": "一次问答：返回完整 committed view（答案只是其中一个字段）"
      }
    },
    "/v1/status": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "description": "error"
          }
        },
        "summary": "可选组件启停（分类器可见面）"
      }
    }
  }
}